[HW2] Tried to add command line arg
Tried to add command line arg to read in obj files. It somewhat works if
the texture coordinates and normals pushback calls in
VulkanLaunchpad.cpp are commented out (lines 2216 ajd 2217). I think it
has to do with the actual pos and normal values being null. Wait on this
to see if TA has anything to say.

Aside from this, ported teapot and icosahedron out to StarterShapes.txt
for now. Included only the icosahedron in objectCreateGeometryAndBuffers
for now.

Next step is to work on arcball camera, disregard the obj files and just
use the icosahedron for now.