Fix specular light and improve moon skin
Specular light is correct now. Look at the git diff, but my normal was wrong
since it took center into account. Haysn didn't like how the moon skin was
oriented so added an offset to "fix" that. Need to make the side of the moon
with the massive crater face earth for the final submission